PIL max pixels
Working with larger areas and smaller zoom levels can result in an error from PIL: PIL.Image.DecompressionBombError : Image size (x pixels) exceeds limit of 178956970 pixels, could be decompression bomb DOS attack.
That's excepted, regular images arent supposed to have so many pixels, but when working with GeoTIFF, that's completely normal.

Anyways, to make the script work with larger areas, a single line of code was needed. Tested it out by myself with a resulting size of 933120000 pixels and it worked well.